code polished
